♻️ refactor(server): split a2a/http.rs by resource
3106-line single file split into 10 modules: error, types, common, conversion, agent_card, task, message, push_config, tests. Largest module now 635 lines (task.rs).
