+++ crates/awaken-server/src/routes.rs
@@ -19,7 +19,7 @@ use crate::config_routes::config_routes;
 use crate::http_run::wire_sse_relay;
 use crate::http_sse::{sse_body_stream, sse_response};
 use crate::mailbox::{ACTIVE_RUN_CONFLICT_MESSAGE, MailboxDispatchStatus, MailboxError};
-use crate::protocols::a2a::http::a2a_routes;
+use crate::protocols::a2a::a2a_routes;
 use crate::protocols::ag_ui::http::ag_ui_routes;
 use crate::protocols::ai_sdk_v6::http::ai_sdk_routes;
 use crate::protocols::mcp::http::mcp_routes;
